pyproject.toml
[tool.poetry]
name = "storitaimu"
version = "0.1.0"
description = ""
authors = ["Nacho114 <17376073+Nacho114@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
llama-index = "^0.12.18"
python-dotenv = "^1.0.1"
llama-index-readers-whisper = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
